Added env wrapper logic
