+++ hive/configs/atari/ppo.yml
@@ -41,10 +41,10 @@ agent:
         gae_lambda: .95
     discount_rate: .99
     grad_clip: .5
-    clip_coef: .1
-    ent_coef: .0
-    clip_vloss: True
-    vf_coef: .5
+    clip_coefficient: .1
+    entropy_coefficient: .0
+    clip_value_loss: True
+    value_fn_coefficient: .5
     transitions_per_update: 4096
     num_epochs_per_update: 4
     normalize_advantages: True
